python/requirements.txt
pyserial
pyautogui
pycaw
numpy
selenium
threading